feat: 添加 Windows 命令桥接功能，支持通过 HTTP 接口执行命令

+++ src/win/winapi.rs
@@ -244,7 +244,7 @@ pub fn patch_dpi_changed_message(hwnd: HWND) {
         return;
     }
     unsafe {
-        let dpi = GetDpiForWindow(hwnd) as u32;
+        let dpi = GetDpiForWindow(hwnd);
         if dpi == 0 {
             return;
         }
